fix: Fix typo in User.findOne method call
The code change fixes a typo in the User.findOne method call in the login function of the auth controller. The typo caused the function to throw an error when trying to find a user by email. This commit corrects the typo, ensuring that the user is properly retrieved from the database.